requirements.txt
Django>=1.3
dj-database-url
south
requests
django-jsonfield
git+http://github.com/lawlesst/py360link.git@sorting-db
git+git://github.com/lawlesst/bibjsontools.git
